# Optimize pkeys and indices by dropping is_current