--- a/discovery-provider/ddl/migrations/0031_drop_current_indices.sql
+++ b/discovery-provider/ddl/migrations/0031_drop_current_indices.sql
@@ -1,0 +1,3 @@
+begin;
+
+commit;