--- a/discovery-provider/integration_tests/queries/test_undisbursed_challeges.py
+++ b/discovery-provider/integration_tests/queries/test_undisbursed_challeges.py
@@ -38,23 +38,6 @@ def setup_challenges(app):
             ),
         ]
 
-        non_current_users = [
-            User(
-                blockhash=hex(99),
-                blocknumber=99,
-                txhash=f"xyz{i}",
-                user_id=i,
-                is_current=False,
-                handle=f"TestHandle{i}",
-                handle_lc=f"testhandle{i}",
-                wallet=f"0x{i}",
-                is_verified=False,
-                name=f"test_name{i}",
-                created_at=datetime.now(),
-                updated_at=datetime.now(),
-            )
-            for i in range(7)
-        ]
         users = [
             User(
                 blockhash=hex(99),
@@ -130,7 +113,6 @@ def setup_challenges(app):
         with db.scoped_session() as session:
             session.add_all(challenges)
             session.flush()
-            session.add_all(non_current_users)
             session.add_all(users)
             session.add_all(user_bank_accounts)
             session.add_all(user_challenges)