### Description
is_current is still referenced from some pkeys and indices even though all rows are is_current true. 

### How Has This Been Tested?
Tested on sandbox. Confirmed migration made the right modifications and indexing is working.

Before
```
audius_discovery=> explain analyze  select * from users where user_id = 1;
                                                          QUERY PLAN                                                           
-------------------------------------------------------------------------------------------------------------------------------
 Index Scan using users_new_pkey on users  (cost=0.56..192059.25 rows=1 width=781) (actual time=0.013..131.900 rows=1 loops=1)
   Index Cond: (user_id = 1)
 Planning Time: 0.102 ms
 Execution Time: 131.931 ms
(4 rows)


audius_discovery=> SELECT
    pg_size_pretty(pg_database_size(current_database())) AS database_size;
 database_size 
---------------
 117 GB

```

After
```
audius_discovery=> explain analyze  select * from users where user_id = 1;
                                                     QUERY PLAN                                                     
--------------------------------------------------------------------------------------------------------------------
 Index Scan using users_pkey on users  (cost=0.56..8.57 rows=1 width=781) (actual time=0.018..0.019 rows=1 loops=1)
   Index Cond: (user_id = 1)
 Planning Time: 0.143 ms
 Execution Time: 0.042 ms


audius_discovery=> SELECT
    pg_size_pretty(pg_database_size(current_database())) AS database_size;
 database_size 
---------------
 106 GB

```


_Please describe the tests that you ran to verify your changes. Provide repro instructions & any configuration._